requirements/dev.txt
-r base.txt
ipdb
ipython
django-debug-toolbar
django-extensions
black
django-browser-reload